drive-assist-api-docker/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.11
Django==4.2.16
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fiona==1.10.1
fonttools==4.54.1
geographiclib==2.0
geopandas==0.14.4
geopy==2.4.1
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
matplotlib==3.9.2
networkx==3.2.1
numpy==1.26.4
osmnx==1.9.4
packaging==24.1
pandas==2.2.3
pillow==11.0.0
PuLP==2.9.0
PyJWT==2.9.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
zipp==3.20.2
